podvm-builder: use submodule to get the cloud-api-adaptor sources
We can't run "git clone" in network isolation, so for our downstream
builds, we have to get the sources differently.

This commit uses a submodule to make the cloud-api-adaptor sources available
during build, using a simple "COPY" command in the Dockerfile to get them
into the image.

I am also modifying the Dockerfile for the operator, as this submodule
conflicts with the way we run "controller-gen" during the operator's build.
Because we use "go run" for it, go will look up all dependencies, and find
conflicting information from the submodule.
To avoid that, this commit removes the podvm-builder sources from the files
we copy to the build container before the build.

Signed-off-by: Julien Ropé <jrope@redhat.com>